+++ clients/admin-ui/src/features/common/custom-fields/CustomFieldsList.tsx
@@ -1,9 +1,14 @@
 import { Center, Divider, Flex, Spinner, Text } from "@fidesui/react";
+import { Field, FieldInputProps, useFormikContext } from "formik";
+import { useCallback, useEffect, useState } from "react";
 
-import { ResourceTypes } from "~/types/api";
+import { useAppDispatch } from "~/app/hooks";
+import { plusApi } from "~/features/plus/plus.slice";
+import { AllowedTypes, ResourceTypes } from "~/types/api";
 
-import { CustomFieldSelector } from "./CustomFieldSelector";
+import { CustomSelect } from "../form/inputs";
 import { CustomFieldsModal } from "./CustomFieldsModal";
+import { filterWithId } from "./helpers";
 import { useCustomFields } from "./hooks";
 
 type CustomFieldsListProps = {
@@ -15,8 +20,11 @@ export const CustomFieldsList = ({
   resourceFidesKey,
   resourceType,
 }: CustomFieldsListProps) => {
+  const dispatch = useAppDispatch();
+  const [selectedKey, setSelectedKey] = useState<string | undefined>();
+  const { isSubmitting, setValues, values } = useFormikContext();
+
   const {
-    definitionIdToCustomField,
     idToAllowListWithOptions,
     idToCustomFieldDefinition,
     isEnabled,
@@ -26,6 +34,50 @@ export const CustomFieldsList = ({
     resourceFidesKey,
     resourceType,
   });
+
+  const handleReload = useCallback(() => {
+    setValues({ ...(values as any) });
+  }, [setValues, values]);
+
+  useEffect(() => {
+    if (!isEnabled || !resourceFidesKey) {
+      return;
+    }
+    // Append the custom field selections to the Formik form
+    if (selectedKey !== resourceFidesKey) {
+      setSelectedKey(resourceFidesKey);
+      dispatch(
+        plusApi.endpoints.getCustomFieldsForResource.initiate(resourceFidesKey)
+      ).then((response) => {
+        const data = new Map(
+          filterWithId(response.data).map((fd) => [
+            fd.custom_field_definition_id,
+            fd,
+          ])
+        );
+        const formValues: Record<string, string | string[]> = {};
+        data.forEach((value, key) => {
+          formValues[`${key}`] = value.value.toString();
+        });
+        setValues(
+          {
+            ...(values as any),
+            ...{ definitionIdToCustomFieldValue: formValues },
+          },
+          false
+        );
+      });
+    }
+  }, [dispatch, isEnabled, resourceFidesKey, selectedKey, setValues, values]);
+
+  useEffect(() => {
+    // Re-render the custom field selections after a Formik form submission
+    if (isSubmitting) {
+      setTimeout(() => {
+        handleReload();
+      });
+    }
+  }, [handleReload, isSubmitting, values]);
 
   if (!isEnabled) {
     return null;
@@ -49,43 +101,58 @@ export const CustomFieldsList = ({
             Custom fields
           </Text>
         </Flex>
-        <CustomFieldsModal resourceType={resourceType} />
+        <CustomFieldsModal reload={handleReload} resourceType={resourceType} />
         {isLoading ? (
           <Center>
             <Spinner />
           </Center>
         ) : (
-          <Flex flexDirection="column" gap="12px" paddingBottom="24px">
-            {sortedCustomFieldDefinitionIds.map((definitionId) => {
-              const customFieldDefinition =
-                idToCustomFieldDefinition.get(definitionId);
-              if (!customFieldDefinition) {
-                // This should never happen.
-                return null;
-              }
+          sortedCustomFieldDefinitionIds.length > 0 && (
+            <Flex flexDirection="column" gap="12px" paddingBottom="24px">
+              {sortedCustomFieldDefinitionIds.map((definitionId) => {
+                const customFieldDefinition =
+                  idToCustomFieldDefinition.get(definitionId);
+                if (!customFieldDefinition) {
+                  // This should never happen.
+                  return null;
+                }
 
-              const allowList = idToAllowListWithOptions.get(
-                customFieldDefinition.allow_list_id!
-              );
-              if (!allowList) {
-                // This would only happen if the field definitions load before
-                // the allow list data.
-                return null;
-              }
+                const allowList = idToAllowListWithOptions.get(
+                  customFieldDefinition.allow_list_id!
+                );
+                if (!allowList) {
+                  // This would only happen if the field definitions load before
+                  // the allow list data.
+                  return null;
+                }
 
-              // This will be undefined when no value has been saved for this field.
-              const customField = definitionIdToCustomField.get(definitionId);
+                const { options } = allowList;
+                const name = `definitionIdToCustomFieldValue.${customFieldDefinition.id}`;
 
-              return (
-                <CustomFieldSelector
-                  allowList={allowList}
-                  customField={customField}
-                  customFieldDefinition={customFieldDefinition}
-                  key={`${definitionId}-${customField?.id}`}
-                />
-              );
-            })}
-          </Flex>
+                return (
+                  <Field key={definitionId} name={name}>
+                    {({
+                      field,
+                    }: {
+                      field: FieldInputProps<string | string[]>;
+                    }) => (
+                      <CustomSelect
+                        {...field}
+                        isClearable
+                        isMulti={
+                          customFieldDefinition.field_type !==
+                          AllowedTypes.STRING
+                        }
+                        label={customFieldDefinition.name}
+                        options={options}
+                        tooltip={customFieldDefinition.description}
+                      />
+                    )}
+                  </Field>
+                );
+              })}
+            </Flex>
+          )
         )}
       </Flex>
     </Flex>