+++ tests/ctl/conftest.py
@@ -19,7 +19,10 @@ orig_requests_delete = requests.delete
 
 @pytest.fixture(scope="session")
 def monkeysession():
-    """monkeypatch fixture at the session level instead of the function level"""
+    """
+    Monkeypatch at the session level instead of the function level.
+    Automatically undoes the monkeypatching when the session finishes.
+    """
     mpatch = MonkeyPatch()
     yield mpatch
     mpatch.undo()
@@ -27,10 +30,11 @@ def monkeysession():
 
 @pytest.fixture(autouse=True, scope="session")
 def monkeypatch_requests(test_client, monkeysession) -> None:
-    """The requests library makes requests against the running webserver
-    which talks to the application db.  This monkeypatching operation
-    makes `requests` calls from src/fides/core/api.py in a test
-    context talk to the test db instead"""
+    """
+    Some places within the application, for example `fides.core.api`, use the `requests`
+    library to interact with the webserver. This fixture patches those `requests` calls
+    so that all of those tests instead interact with the test instance.
+    """
     monkeysession.setattr(requests, "get", test_client.get)
     monkeysession.setattr(requests, "post", test_client.post)
     monkeysession.setattr(requests, "put", test_client.put)